kafka-app/producer/requirements.txt
fastapi
uvicorn
confluent-kafka
pydantic
pydantic-settings
python-dotenv
requests
debugpy